+++ Changelog.md
@@ -4,14 +4,16 @@
 
 - Add roll button on Eternity Shards to roll Reality against the Tapping Difficulty of the Shard.
 - Add a button in SK sheets and Eternity Shards to reset the number of possibilities at the start of the Act.
+  - Hold SHIFT on the SK sheet to automatically reset used possibilities on owned Eternity Shards as well.
 - Allow number of possibilities per Act for SKs to be configured with `CONFIG.torgeternity.possibilitiesPerAct`
   - Active Effects can modify the number per act by modifying the property `system.other.possibilities.perAct`
+  - Any existing AE that modifies `system.other.possibilities` will be automatically changed to modify `system.other.possibilities.perAct` (e.g. Prodigy perk)
 
 ### Translations
 
 - `torgeternity.chatText.tapping`: "Tapping Test"
 - `torgeternity.iconeTitles.tappingRoll`: "Tapping Roll"
-- `torgeternity.iconeTitles.resetPoss`: "Reset Possibilities"
+- `torgeternity.iconeTitles.resetPoss`: "Reset Possibilities (hold SHIFT to include Eternity Shards)"
 
 ## 13.18.2 - Fix Custom Skill Name in chat message